requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
cohere==5.12.0
dataclasses-json==0.6.7
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
faiss-cpu==1.9.0.post1
fastavro==1.9.7
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-cohere==0.3.3
langchain-community==0.3.8
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.147
Markdown==3.7
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.2
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pillow==11.0.0
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.2
style==1.1.0
tabulate==0.9.0
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
update==0.0.1
urllib3==2.2.3
yarl==1.18.0
